refactor: Token 설정 관리하는 JwtTokenProperties 클래스 분리
