+++ src/main/java/roomescape/infrastructure/auth/JwtTokenProvider.java
@@ -5,48 +5,41 @@ import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import java.util.Date;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import roomescape.application.TokenProvider;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
 
 @Component
-@ConfigurationProperties(prefix = "security.jwt.token")
 public class JwtTokenProvider implements TokenProvider {
-    private String secretKey;
-    private long validityInMilliseconds;
+    private final JwtTokenProperties jwtTokenProperties;
+
+    public JwtTokenProvider(JwtTokenProperties jwtTokenProperties) {
+        this.jwtTokenProperties = jwtTokenProperties;
+    }
 
     public String createToken(String payload) {
         Claims claims = Jwts.claims().setSubject(payload);
         Date now = new Date();
-        Date validity = new Date(now.getTime() + validityInMilliseconds);
+        Date validity = new Date(now.getTime() + jwtTokenProperties.getValidityInMilliseconds());
 
         return Jwts.builder()
                 .setClaims(claims)
                 .setIssuedAt(now)
                 .setExpiration(validity)
-                .signWith(SignatureAlgorithm.HS256, secretKey)
+                .signWith(SignatureAlgorithm.HS256, jwtTokenProperties.getSecretKey())
                 .compact();
     }
 
     public String getPayload(String token) {
         try {
             return Jwts.parser()
-                    .setSigningKey(secretKey)
+                    .setSigningKey(jwtTokenProperties.getSecretKey())
                     .parseClaimsJws(token)
                     .getBody()
                     .getSubject();
         } catch (ExpiredJwtException e) {
             throw new RoomescapeException(RoomescapeErrorCode.TOKEN_EXPIRED);
         }
-    }
-
-    public void setSecretKey(String secretKey) {
-        this.secretKey = secretKey;
-    }
-
-    public void setValidityInMilliseconds(long validityInMilliseconds) {
-        this.validityInMilliseconds = validityInMilliseconds;
     }
 }
